requirements.txt
selenium==4.11.2
requests
python-dotenv
webdriver-manager
